Added files to plot the redback plots in ToO format
The notebook simulates the Metzger Kilonova model with suitable parameters using manual sky pointings files generated for the Golden Strategy for the 3 different cases as per the ToO report. This notebook was used to generate the Faint Kilonova plots used in the Arxiv/published format of the Rubin ToO 2024 workshop report.